requirements/localhost.txt
# requirements/dev.txt
-r common.txt
django-debug-toolbar==1.2.1
